% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat_summary.R
\name{qstat_summary}
\alias{qstat_summary}
\title{General Statistical Summary}
\usage{
qstat_summary(x, removeNA = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector for which the statistical summary is to be computed. 'x' cannot be a list, and the input vector should not contain NA values unless the 'removeNA' argument is specified to be TRUE.}

\item{removeNA}{a logical argument for the handling of NA values in the input vector. The function will display an error message if NA values are contained in the input vector (default set to 'removeNA = FALSE'); specify 'removeNA = TRUE' to remove NA values from the input vector before summary statistics are calculated.}

\item{verbose}{a logical argument allowing the user to specify whether they would like feedback on the workings of the function, while the function is running on the input vector. Default is set to 'verbose = FALSE', so that no messages are displayed, if this argument has not been explicitly specified.}

\item{...}{additional arguments that can be passed on to the mean, sd, max, min and quantile functions within the main function.}
}
\value{
A tibble consisting of seven columns and a single row, where the entry for each column is a calculated general statistic (mean, standard deviation, extrema and quartiles) for the input numeric vector.
}
\description{
General Statistical Summary
}
\examples{
qstat_summary(c(1.2, 5.6, 11.2, 4.9, 3.8))
}
