% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:summary_test.R
\name{summary_test}
\alias{summary_test}
\title{Load a Summary}
\usage{
summary_test(variable)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function calculates the statistics of a dataset as the mean, variance, minimum, maximum ...
contains the rownames and the subsequent columns are the sample identifiers.
}
